Some file cleanup and re-enable NPM build attempt
