+++ mod.ts
@@ -1,3 +1,3 @@
 export * from "./src/atoms/Icon.tsx";
-export * from "./src/sprites/IconSet.tsx";
+export * from "./src/sprites/_exports.tsx";
 export * from "./src/plugin/IconSet.plugin.tsx";
