+++ src/sprites/IconSetConfig.tsx
@@ -1,0 +1,3 @@
+export interface IconSetConfig {
+  IconMap: Record<string, string | URL>;
+}
