+++ src/sprites/IconSetGenerateConfig.tsx
@@ -1,0 +1,13 @@
+import { IconSetConfig } from "./IconSetConfig.tsx";
+
+export interface IconSetGenerateConfig {
+  Exports?: boolean;
+
+  Imports?: string;
+
+  OutputDirectory?: string;
+
+  IconSet: IconSetConfig;
+
+  SpriteSheet: string;
+}
