+++ src/sprites/_exports.tsx
@@ -1,0 +1,5 @@
+export * from "./component.utils.tsx";
+export * from "./IconSet.tsx";
+export * from "./IconSetConfig.tsx";
+export * from "./IconSetGenerateConfig.tsx";
+export * from "./sheet.utils.tsx";
