+++ tests/sprites/IconSet.tests.tsx
@@ -5,12 +5,9 @@ import {
   render,
   ServeHandlerInfo,
 } from "../tests.deps.ts";
-import {
-  IconSet,
-  IconSetConfig,
-  useFileIconSet,
-  useIconSetComponents,
-} from "../../src/sprites/IconSet.tsx";
+import { IconSet } from "../../src/sprites/IconSet.tsx";
+import { IconSetConfig } from "../../src/sprites/IconSetConfig.tsx";
+import { useFileIconSet } from "../../src/sprites/sheet.utils.tsx";
 // import manifest from "../../fresh.gen.ts";
 
 // const CONN_INFO: ServeHandlerInfo = {
